Refactor CI tests to resolve port conflict
The CI pipeline was failing due to a port conflict. The GitHub Actions workflow was starting a PostgreSQL service, and the `pytest-docker` plugin was simultaneously trying to start another one on the same port.

This commit refactors the tests to resolve this conflict. The `pytest-docker` functionality has been removed from the tests. Instead, the tests now rely on the service containers provided by the GitHub Actions workflow.

The changes include:
- Removing the `docker_services` fixture from the integration tests.
- Centralizing the service health check fixtures into `tests/conftest.py`.
- Ensuring the tests connect to the services started by the CI workflow.

This new approach is more robust and aligns with best practices for using service containers in GitHub Actions. It should provide a stable and reliable CI pipeline.
